+++ whois-utils.go
@@ -15,23 +15,17 @@ import (
     "io/ioutil"
 )
 
-
-func ReadFile(file string) (result string, err error) {
-    tmp_result, err := ioutil.ReadFile(file)
+func ReadFile(file string) (str string, err error) {
+    tmpResult, err := ioutil.ReadFile(file)
     if err != nil {
         return
     }
-
-    result = string(tmp_result)
-    return
+    return string(tmpResult), nil
 }
 
-
-func WriteFile(file string, data string) (err error) {
-    err = ioutil.WriteFile(file, []byte(data), 0644)
-    return
+func WriteFile(file string, data string) error {
+    return ioutil.WriteFile(file, []byte(data), 0644)
 }
-
 
 func IsNotFound(data string) (result bool) {
     data = strings.ToLower(data)
@@ -40,7 +34,6 @@ func IsNotFound(data string) (result bool) {
         strings.Contains(data, "no entries found") || strings.Contains(data, "no data found") ||
         strings.Contains(data, "not registered") || strings.Contains(data, "is free")
 }
-
 
 func TransferName(name string) (string) {
     name = strings.ToLower(name)
@@ -143,21 +136,18 @@ func TransferName(name string) (string) {
     return name
 }
 
-
 func RemoveDuplicateField(data string) string {
-    new_fields := []string{}
-    fields := strings.Split(data, ",")
-    for _, v := range fields {
+    var newFields []string
+    for _, v := range strings.Split(data, ",") {
         if v == "" {
             continue
         }
-        if !StringInArray(new_fields, v) {
-            new_fields = append(new_fields, v)
+        if !StringInArray(newFields, v) {
+            newFields = append(newFields, v)
         }
     }
-    return strings.Join(new_fields, ",")
+    return strings.Join(newFields, ",")
 }
-
 
 func StringInArray(array []string, find string) bool {
     for _, v := range array {
@@ -167,7 +157,6 @@ func StringInArray(array []string, find string) bool {
     }
     return false
 }
-
 
 func FixNameServers(nservers string) string {
     servers := strings.Split(nservers, ",")
